Update run scripts with correct data location and Python version
Pointing to Prasanna's checked-out copy of "sotto-voce-corpus".  This repository
seems to have been removed from Github, so we'll want to hang on to this copy.

Switched to use version of Python in the path.  That way, you can activate a
virtual environment that has everything set up prior to running and use its
Python.